# Rebuilding _StringCore [Not ready for merge]