--- a/stdlib/public/SDK/CoreGraphics/CGFloat.swift.gyb
+++ b/stdlib/public/SDK/CoreGraphics/CGFloat.swift.gyb
@@ -57,10 +57,12 @@ public struct CGFloat {
   }
 
 % for src_ty in all_integer_types(word_bits):
+%   if src_ty.access == 'public':
   @_transparent public init(_ value: ${src_ty.stdlib_name}) {
     self.native = NativeType(value)
   }
 
+%   end
 % end
 
   /// The native value.
@@ -390,6 +392,7 @@ extension CGFloat : Hashable {
 }
 
 % for dst_ty in all_integer_types(word_bits):
+%   if dst_ty.access == 'public':
 
 extension ${dst_ty.stdlib_name} {
   @_transparent
@@ -398,6 +401,7 @@ extension ${dst_ty.stdlib_name} {
   }
 }
 
+%   end
 % end