--- a/stdlib/public/SDK/CoreGraphics/CGFloat.swift.gyb
+++ b/stdlib/public/SDK/CoreGraphics/CGFloat.swift.gyb
@@ -56,13 +56,13 @@ public struct CGFloat {
     self.native = value.native
   }
 
-% for src_ty in all_integer_types(word_bits):
-%   if src_ty.access == 'public':
+% for src_ty in (
+%   x for x in all_integer_types(word_bits) 
+%   if not x.stdlib_name.startswith('_')):
   @_transparent public init(_ value: ${src_ty.stdlib_name}) {
     self.native = NativeType(value)
   }
 
-%   end
 % end
 
   /// The native value.
@@ -392,7 +392,6 @@ extension CGFloat : Hashable {
 }
 
 % for dst_ty in all_integer_types(word_bits):
-%   if dst_ty.access == 'public':
 
 extension ${dst_ty.stdlib_name} {
   @_transparent
@@ -401,7 +400,6 @@ extension ${dst_ty.stdlib_name} {
   }
 }
 
-%   end
 % end