… stdlib/public/core/Builtin.swift
… stdlib/public/core/CMakeLists.txt
… stdlib/public/core/Collection.swift
… stdlib/public/core/ContiguousArrayBuffer.swift
… stdlib/public/core/ContiguouslyStored.swift
… stdlib/public/core/Counted.swift.gyb
… stdlib/public/core/FloatingPoint.swift.gyb
… stdlib/public/core/FloatingPointTypes.swift.gyb
… stdlib/public/core/ForwardCollectionAPIs.swift.gyb
… stdlib/public/core/ForwardingWrapper.swift.gyb
… stdlib/public/core/Function.swift
… stdlib/public/core/GroupInfo.json
… stdlib/public/core/Integers.swift.gyb
… stdlib/public/core/LazyCollection.swift.gyb
… stdlib/public/core/Map.swift.gyb
… stdlib/public/core/MutableCollection.swift
… stdlib/public/core/RangeReplaceableCollection.swift.gyb
… stdlib/public/core/SentinelCollection.swift
… stdlib/public/core/Sequence.swift
… stdlib/public/core/SequenceAlgorithms.swift.gyb
… stdlib/public/core/SequenceWrapper.swift
… stdlib/public/core/Slice.swift.gyb
… stdlib/public/core/StringBridge.swift
… stdlib/public/core/StringCore2.swift
… stdlib/public/core/StringStorage.swift
… stdlib/public/core/UnsafeBufferPointer.swift.gyb
… stdlib/public/stubs/FoundationHelpers.mm
… stdlib/public/stubs/GlobalObjects.cpp
… test/IRGen/literals.sil
… test/Prototypes/s.swift