A replacement for _StringCore, that also lays the groundwork for maintaining NUL termination in our buffers.

In case someone wants to carry this forward, the [best commit in this branch](https://github.com/dabrahams/swift/commit/2c2c7b50c1ff71b20470cee66982d41a99ae26ff) produces the following results: 

Test|Time|Variance
----|----|--------
`lexicographicalComparison_old()`|                   1103|      ±12
`lexicographicalComparison_new()`|                   760|       ±2
`initFromArray_old()`|                               1479|      ±21
`initFromArray_new()`|                               760|       ±6
`appendManyTinyASCIIFragments_ToASCII_old()`|        566|       ±11
`appendManyTinyASCIIFragments_ToASCII_new()`|        608|       ±4
`appendManyTinyFragmentsOfBothWidths_old()`|         706|       ±11
`appendManyTinyFragmentsOfBothWidths_new()`|         1177|      ±2
`appendManyLongASCII_ToUTF16_old()`|                 991|       ±5
`appendManyLongASCII_ToUTF16_new()`|                 967|       ±14
`appendFewTinyASCIIFragments_ToASCII_old()`|         1023|      ±2
`appendFewTinyASCIIFragments_ToASCII_new()`|         614|       ±5
`appendFewTinyFragmentsOfBothWidths_old()`|          2432|      ±16
`appendFewTinyFragmentsOfBothWidths_new()`|          1773|      ±5
`appendOneLongASCII_ToUTF16_old()`|                  271|       ±3
`appendOneLongASCII_ToUTF16_new()`|                  138|       ±2

This shows we beat the current `_StringCore` handily in the our official benchmarks' string builder tests, and in a few other tests besides.  Our current string builder benchmarks are notably lacking coverage for some important cases (see `appendManyTinyFragmentsOfBothWidths_new()` above, which is harmed by ARC traffic that won't seem to quit), and we lag in at least a few of those, but never by more than 50%.

That said, the [final commit in this branch](https://github.com/dabrahams/swift/commit/6f820538e1d2dcaff64a702a056f1b93d4c849c9), which moved the code of the new prototype into the standard library, produced massive slowdowns in several benchmarks:

Test|Time Before 6f820538e1d2|Time After 6f820538e1d2
----|-----------------------|----------------------
appendManyTinyASCIIFragments_ToASCII_new()|        608|       2101
appendManyTinyFragmentsOfBothWidths_new()|         1177|      2218
appendFewTinyASCIIFragments_ToASCII_new()|         614|       2161
appendFewTinyFragmentsOfBothWidths_new()|          1773|      3587

I believe I've pushed this work as far as I can take it from the library side and now it's up to the optimizer.