@dabrahams 2017-07-06
@aschwaighofer @milseman [This branch](/dabrahams/swift/tree/builtin-int128) tries to use Builtin.Int128 directly to avoid some of the overheads apparently introduced by the internal `UInt128`.  It got maybe 5-10% back on the latter three slowdowns and made no difference to appendManyTinyASCIIFragments_ToASCII